Decorate the original type, not replace it
The current implementation will clobber the underlying datatype. This
means that we lose any casting behavior that it has post-serialization,
and any internal methods aren't delegated.

This specifically caused issues in Shopify core because in 4.2, Active
Record column objects have access to type objects for each attribute,
and delegate internal methods like `type` to it. In Rails 5 columns know
nothing about types, and have a metadata object instead, and this was
leading to cache keys for IDC having a mismatch.

`decorate_attribute_type` is not part of the public external API, but
I super promise that it's stable and safe to rely on.