Support Rails 5.2+
As of Rails 5.2, `attribute_will_change!` no longer works for
non-attributes. This re-implements the required methods of `Dirty`
manually, similar to what was originally done by this gem.

Note: I don't think it's appropriate to override `changes`. It's
reasonable to expect that `model.changes` will never include keys that
aren't also keys in `model.attributes` or `model.attribute_names`.
I think it would be more appropriate to support the `from:` and `to:`
options in `attribute_changed?` instead.